Refactored Hypnos to support different interrupt triggers and different sleep triggers

+++ src/Sensors/I2C/Loom_ADS1115/Loom_ADS1115.h
@@ -44,13 +44,13 @@ class Loom_ADS1115 : public Module{
          * Get the value in the analog table according to analog number not INDEX
          * @param pin Pin to get data from (1-4)
          */ 
-        int16_t getAnalog(int pin) { return analogData[pin-1]; };
+        int getAnalog(int pin) { return analogData[pin-1]; };
 
         /**
          * Get the value in the diff table according to diff number not INDEX
          * @param pin Pin to get data from (1-2)
          */ 
-        int16_t getDiff(int pin) { return analogData[pin-1]; };
+        int getDiff(int pin) { return analogData[pin-1]; };
     
     private:
         Manager* manInst;                                           // Instance of the manager
@@ -61,8 +61,8 @@ class Loom_ADS1115 : public Module{
         bool enableAnalog;                                          // Read from the analog pins
         bool enableDiff;                                            // Read differentials
 
-        int16_t analogData[4];                                      // Stores the analog ADS1115 data
-        int16_t diffData[2];                                        // Stores the differential data from the sensor
+        int analogData[4];                                          // Stores the analog ADS1115 data
+        int diffData[2];                                            // Stores the differential data from the sensor
 
         bool initialized = true;
